@@ -1,13 +1,13 @@
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
-    id("com.google.protobuf") version "0.8.8"
+    id("com.google.protobuf") version "0.8.16"
     id("org.jetbrains.kotlin.jvm")
     id("idea")
 }
 
 ext {
-    protobufVersion = '3.6.1'
+    protobufVersion = '4.0.0-rc-2'
 }
 
 protobuf {
@@ -31,6 +31,6 @@ dependencies {
 tasks.withType(KotlinCompile) {
     kotlinOptions {
         allWarningsAsErrors = true
-        jvmTarget = "1.8"
+        jvmTarget = "15"
     }
 }
