@@ -4,8 +4,8 @@ kotlin {
     sourceSets {
         all {
             languageSettings {
-                languageVersion = '1.4'
-                apiVersion = '1.4'
+                languageVersion = '1.5'
+                apiVersion = '1.5'
                 useExperimentalAnnotation('kotlin.ExperimentalUnsignedTypes')
                 progressiveMode = true
             }
