@@ -4,7 +4,7 @@ kotlin {
         compilations.all {
             kotlinOptions {
                 allWarningsAsErrors = true
-                jvmTarget = '1.8'
+                jvmTarget = '15'
                 freeCompilerArgs += "-Xopt-in=kotlin.RequiresOptIn"
             }
         }
